+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Enhancements
 - Added new publisher values for Cboe Titanium Cboe Global Indices Feed
 - Added `StatType` variants `MwcbLevel1`, `MwcbLevel2`, and `MwcbLevel3`
+- Added `StatType` variants `AuctionCollarReferencePrice`, `AuctionCollarUpperPrice`,
+  and `AuctionCollarLowerPrice`
 
 ### Breaking changes
 - Renamed the following Venue, Dataset, and Publisher: